+++ src/OpenLoco/src/GameCommands/Vehicles/RenameVehicle.cpp
@@ -33,10 +33,10 @@ namespace OpenLoco::GameCommands
     {
         setExpenditureType(ExpenditureType::TrainRunningCosts);
 
-        static loco_global<EntityId, 0x0113621D> _headId_113621D;
+        static EntityId targetVehicleHeadId{}; // 0x0113621D
         if (args.i == 1)
         {
-            _headId_113621D = args.head;
+            targetVehicleHeadId = args.head;
         }
 
         static char staticRenameBuffer[37]{};
@@ -53,7 +53,7 @@ namespace OpenLoco::GameCommands
             return 0;
         }
 
-        EntityId vehicleHeadId = _headId_113621D;
+        EntityId vehicleHeadId = targetVehicleHeadId;
         Vehicles::VehicleHead* vehicleHead = EntityManager::get<Vehicles::VehicleHead>(vehicleHeadId);
 
         if (vehicleHead == nullptr)