+++ src/OpenLoco/src/OpenLoco.h
@@ -22,6 +22,7 @@ namespace OpenLoco
     void simulateGame(const fs::path& path, int32_t ticks);
 
     void sub_431695(uint16_t var_F253A0);
+    uint16_t getTimeSinceLastTick();
     int main(std::vector<std::string>&& argv);
     bool promptTickLoop(std::function<bool()> tickAction);
     [[noreturn]] void exitCleanly();