+++ src/OpenLoco/src/Ui/WindowManager.h
@@ -47,6 +47,9 @@ namespace OpenLoco::Ui::WindowManager
 
     WindowType getCurrentModalType();
     void setCurrentModalType(WindowType type);
+
+    void resetThousandthTickCounter();
+
     Window* get(size_t index);
     size_t indexOf(const Window& pWindow);
     size_t count();
@@ -125,7 +128,9 @@ namespace OpenLoco::Ui::Windows
 
     namespace BuildVehicle
     {
-        Window* open(uint32_t vehicle, uint32_t flags);
+        Window* openByVehicleId(EntityId vehicleId);
+        Window* openByType(VehicleType vehicleType);
+        Window* openByVehicleObjectId(uint16_t vehicleObjectId);
         void sub_4B92A5(Ui::Window* window);
     }