+++ src/OpenLoco/src/Ui/Windows/News/News.cpp
@@ -27,15 +27,9 @@
 #include "World/IndustryManager.h"
 #include "World/StationManager.h"
 #include "World/TownManager.h"
-#include <OpenLoco/Interop/Interop.hpp>
-
-using namespace OpenLoco::Interop;
 
 namespace OpenLoco::Ui::Windows::NewsWindow
 {
-    static loco_global<uint32_t, 0x011364EC> _numTrackTypeTabs;
-    static loco_global<int8_t[8], 0x011364F0> _trackTypesForTab;
-
     namespace News1
     {
         static constexpr auto widgets = makeWidgets(
@@ -142,29 +136,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                                 break;
 
                             case MessageItemArgumentType::vehicleTab:
-                                auto vehicleObj = ObjectManager::get<VehicleObject>(itemId);
-                                auto window = Ui::Windows::BuildVehicle::open(static_cast<uint32_t>(vehicleObj->type), (1U << 31));
-                                window->rowHover = itemId;
-                                if (vehicleObj->mode == TransportMode::rail || vehicleObj->mode == TransportMode::road)
-                                {
-                                    if (vehicleObj->trackType != 0xFF)
-                                    {
-                                        for (uint8_t i = 0; i < _numTrackTypeTabs && i < std::size(_trackTypesForTab); ++i)
-                                        {
-                                            if (vehicleObj->trackType == _trackTypesForTab[i])
-                                            {
-                                                window->currentSecondaryTab = i;
-                                                break;
-                                            }
-                                        }
-                                    }
-                                }
-
-                                auto rowHover = window->rowHover;
-
-                                Ui::Windows::BuildVehicle::sub_4B92A5(window);
-
-                                window->rowHover = rowHover;
+                                Ui::Windows::BuildVehicle::openByVehicleObjectId(itemId);
                                 break;
                         }
                     }